Correct audio type in XML feed
Previously this was harded to "ogg", while in fact the only user of this
template at the moment uses "mp3". This means the resulting XML file was
actually wrong. This doesn't matter too much, but it may as well be
correct, by passing along the same format that Polly is requested to
produce.